requirements.txt
google-auth==2.26.2
google-api-python-client==2.114.0
python-dotenv==1.0.0